correction of comments made by Henry
