+++ training-react/tic-tac-toe/src/app/screens/Home/index.js
@@ -1,6 +1,7 @@
 import React, { Fragment } from 'react';
 import { Switch, Route } from 'react-router-dom';
 import ROUTES from '@consts/route';
+import ErrorPage from '@screens/ErrorPage';
 
 import Game from './screens/Game';
 import Profile from './screens/Profile';
@@ -15,6 +16,7 @@ function Home() {
         <Route exact path={ROUTES.HOME} component={HomePage} />
         <Route exact path={ROUTES.GAME} component={Game} />
         <Route exact path={ROUTES.PROFILE} component={Profile} />
+        <Route component={ErrorPage} />
       </Switch>
     </Fragment>
   );
