+++ training-react/tic-tac-toe/src/app/screens/Home/screens/Profile/components/Form/styles.scss
@@ -4,10 +4,6 @@
   width: 100%;
 }
 
-.prueba {
-  width: 50%;
-}
-
 .btn-login {
   background: $blue-indigo;
   color: $white;
@@ -41,7 +37,7 @@
   font-size: 20px;
   margin-top: 15px;
   text-align: center;
-  width: 100%;;
+  width: 100%;
 }
 
 .success-update {
