+++ training-react/tic-tac-toe/src/app/screens/Home/screens/Profile/index.js
@@ -1,6 +1,7 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import actionCreators from '@redux/profile/actions';
+import { infoUserConst } from '@consts/propTypes';
 import PropTypes from 'prop-types';
 
 import Layout from './layout';
@@ -8,8 +9,7 @@ import Layout from './layout';
 class Profile extends Component {
   componentDidMount() {
     const { getUser } = this.props;
-    const idUser = localStorage.getItem('userId');
-    getUser(idUser);
+    getUser();
   }
 
   componentWillUnmount() {
@@ -33,23 +33,18 @@ class Profile extends Component {
 Profile.propTypes = {
   updateUser: PropTypes.func.isRequired,
   getUser: PropTypes.func.isRequired,
-  infoUser: PropTypes.shape({
-    firstname: PropTypes.string,
-    surname: PropTypes.string,
-    username: PropTypes.string,
-    address: PropTypes.string
-  }),
+  clearState: PropTypes.func.isRequired,
   loaded: PropTypes.bool.isRequired,
   msgUpdate: PropTypes.string.isRequired,
   hasError: PropTypes.bool.isRequired,
-  clearState: PropTypes.func
+  infoUser: infoUserConst
 };
 
-const mapStateToProps = state => ({
-  infoUser: state.profile.info,
-  loaded: state.profile.loaded,
-  msgUpdate: state.profile.msgUpdate,
-  hasError: state.profile.hasError
+const mapStateToProps = ({ profile }) => ({
+  infoUser: profile.info,
+  loaded: profile.loaded,
+  msgUpdate: profile.msgUpdate,
+  hasError: profile.hasError
 });
 
 const mapDispatchToProps = dispatch => ({
