+++ training-react/tic-tac-toe/src/app/screens/Home/screens/Profile/layout.js
@@ -1,10 +1,11 @@
 import React from 'react';
 import PropTypes from 'prop-types';
+import { infoUserConst } from '@consts/propTypes';
 
 import Form from './components/Form';
 import styles from './styles.scss';
 
-function Layout({ onSubmit, loaded, infoUser, msgUpdate, hasError }) {
+function Layout({ onSubmit, infoUser, loaded, msgUpdate, hasError }) {
   return (
     <section className={styles.profile}>
       <article className={styles.containerForm}>
@@ -19,14 +20,9 @@ function Layout({ onSubmit, loaded, infoUser, msgUpdate, hasError }) {
 Layout.propTypes = {
   onSubmit: PropTypes.func.isRequired,
   loaded: PropTypes.bool.isRequired,
-  infoUser: PropTypes.shape({
-    firstname: PropTypes.string,
-    surname: PropTypes.string,
-    username: PropTypes.string,
-    address: PropTypes.string
-  }).isRequired,
   msgUpdate: PropTypes.string.isRequired,
-  hasError: PropTypes.bool.isRequired
+  hasError: PropTypes.bool.isRequired,
+  infoUser: infoUserConst
 };
 
 export default Layout;
