+++ training-react/tic-tac-toe/src/app/screens/Login/styles.scss
@@ -28,7 +28,7 @@
 }
 
 .btn-login {
-  background: $blue-indigo;
+  background-color: $blue-indigo;
   color: $white;
   font-size: 14px;
   padding: 15px;
