+++ training-react/tic-tac-toe/src/constants/propTypes.js
@@ -1,0 +1,8 @@
+import PropTypes from 'prop-types';
+
+export const infoUserConst = PropTypes.shape({
+  firstname: PropTypes.string,
+  surname: PropTypes.string,
+  username: PropTypes.string,
+  address: PropTypes.string
+});
